+++ saas/backend/apps/subject/views.py
@@ -14,7 +14,7 @@ from rest_framework import serializers, status
 from rest_framework.response import Response
 from rest_framework.viewsets import GenericViewSet
 
-from backend.account.permissions import role_perm_class, system_access_perm_class
+from backend.account.permissions import role_perm_class
 from backend.account.serializers import AccountRoleSLZ
 from backend.apps.group.audit import GroupMemberDeleteAuditProvider
 from backend.apps.group.models import Group
@@ -128,13 +128,6 @@ class SubjectPolicyViewSet(BizMixin, GenericViewSet):
 
     pagination_class = None  # 去掉 swagger 中的 limit offset 参数
 
-    def get_permissions(self):
-        if self.action in ("list", "destroy"):
-            return super().get_permissions() + [system_access_perm_class("query", "system_id")()]
-        if self.action == "update":
-            return super().get_permissions() + [system_access_perm_class("body", "system_id")()]
-        return super().get_permissions()
-
     @swagger_auto_schema(
         operation_description="Subject 权限列表",
         query_serializer=SystemQuerySLZ(),
@@ -149,6 +142,9 @@ class SubjectPolicyViewSet(BizMixin, GenericViewSet):
         slz.is_valid(raise_exception=True)
 
         system_id = slz.validated_data["system_id"]
+
+        # 校验系统访问权限
+        self.system_biz.validate_system_access(system_id)
 
         policies = self.policy_query_biz.list_by_subject(system_id, subject)
 
@@ -172,6 +168,10 @@ class SubjectPolicyViewSet(BizMixin, GenericViewSet):
         slz.is_valid(raise_exception=True)
 
         system_id = slz.validated_data["system_id"]
+
+        # 校验系统访问权限
+        self.system_biz.validate_system_access(system_id)
+
         ids = slz.validated_data["ids"]
 
         # 为了记录审计日志，需要在删除前查询
@@ -209,6 +209,10 @@ class SubjectPolicyViewSet(BizMixin, GenericViewSet):
 
         # 为避免需要忽略的变量与国际化翻译变量"_"冲突，所以使用"__"
         system_id = self.policy_query_biz.get_policy_system_by_id(subject, policy_id)
+
+        # 校验系统访问权限
+        self.system_biz.validate_system_access(system_id)
+
         update_policy = self.policy_operation_biz.delete_partial(
             system_id,
             subject,
@@ -240,6 +244,10 @@ class SubjectPolicyResourceGroupDeleteViewSet(BizMixin, GenericViewSet):
         # FIXME(tenant): 需要校验 subject 是否是当前租户的用户或部门
 
         system_id = self.policy_query_biz.get_policy_system_by_id(subject, policy_id)
+
+        # 校验系统访问权限
+        self.system_biz.validate_system_access(system_id)
+
         # 删除权限
         update_policy = self.policy_operation_biz.delete_by_resource_group_id(
             system_id, subject, policy_id, resource_group_id
@@ -272,7 +280,6 @@ class SubjectRoleViewSet(BizMixin, GenericViewSet):
 class SubjectTemporaryPolicyViewSet(BizMixin, GenericViewSet):
     permission_classes = [
         role_perm_class(PermissionCodeEnum.MANAGE_ORGANIZATION.value),
-        system_access_perm_class("query", "system_id"),
     ]
 
     pagination_class = None  # 去掉 swagger 中的 limit offset 参数
@@ -288,6 +295,9 @@ class SubjectTemporaryPolicyViewSet(BizMixin, GenericViewSet):
         slz.is_valid(raise_exception=True)
 
         system_id = slz.validated_data["system_id"]
+
+        # 校验系统访问权限
+        self.system_biz.validate_system_access(system_id)
 
         subject = Subject(type=kwargs["subject_type"], id=kwargs["subject_id"])
         # FIXME(tenant): 需要校验 subject 是否是当前租户的用户或部门
@@ -308,6 +318,10 @@ class SubjectTemporaryPolicyViewSet(BizMixin, GenericViewSet):
         slz.is_valid(raise_exception=True)
 
         system_id = slz.validated_data["system_id"]
+
+        # 校验系统访问权限
+        self.system_biz.validate_system_access(system_id)
+
         ids = slz.validated_data["ids"]
         subject = Subject(type=kwargs["subject_type"], id=kwargs["subject_id"])
